refactor(server): split member and application responsibilities
